\name{prost.y}
\docType{data}
\alias{prost.y}
\title{Labels for Tumoral and Normal Prostate Samples}
\description{Labels for a random sample of 25 normal samples and 25 tumour samples from the normalised subset from Singh et al. (2002) data included in the \code{prostate} dataset.}
\usage{data(prostate)}
\format{
a vector of length 50 containing the sample type: 0=normal, 1=tumor.
}
\source{The data are described in Singh et al. (2002).}
\references{ Singh \emph{et al}. (2002). Gene expression correlates of clinincal prostate cancer behavior, \emph{Cancer cell}, 1 (2), 203-209.
 
   Dudoit \emph{et al}. (2002). Comparison of discrimination methods for the classification of tumors using gene expression data, \emph{Journal of the American Statistical Association}, 97 (457), 77-87.
}
\keyword{datasets}
